chore: move row value classes out of row_data
This is in preparation for extracting row merging into a separate class. See https://github.com/googleapis/python-bigtable/pull/628
